# api/swagger.yml
swagger: "2.0"
info:
  description: "This is a Stonks service. You can find out more about us at [https://github.com/veremchukvv/stonks-test](https://github.com/veremchukvv/stonks-test)."
  version: "1.0.0"
  title: "Stonks"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/mit-license.php"
schemes:
  - "http"
basePath: "/"
...
paths:
...
  /api/v1/portfolios/{portfolioId}:
    get:
      tags:
        - "portfolios"
      summary: "Get portfolio deals"
      operationId: "getPortfolioDealsById"
      produces:
        - "application/json"
      parameters:
        - name: "portfolioId"
          in: "path"
          description: "ID of portfolio to return"
          required: true
          type: "integer"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/PortfolioDeals"
        "401":
          description: "Not logined"
        "500":
          description: "Error on get deals"
    delete:
      tags:
        - "portfolios"
      summary: "Delete a portfolio"
      description: ""
      operationId: "deletePortfolio"
      produces:
        - "application/json"
      parameters:
        - name: "portfolioId"
          in: "path"
          description: "Portfolio id to delete"
          required: true
          type: "integer"
      responses:
        "200":
          description: "OK"
        "401":
          description: "Not logined"
        "500":
          description: "Error on deleting portfolio"
...
  /api/v1/deals/{dealId}:
    get:
      tags:
        - "deals"
      summary: "Get information about one deal"
      operationId: "getDeal"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "dealId"
          in: "path"
          description: "ID of deal"
          required: true
          type: "integer"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/Deals"
        "500":
          description: "Error on getting deal"
    post:
      tags:
        - "deals"
      summary: "Close deal"
      operationId: "closeDeal"
      produces:
        - "application/json"
      parameters:
        - name: "dealId"
          in: "path"
          description: "ID of closing deal"
          required: true
          type: "integer"
      responses:
        "200":
          description: "OK"
        "401":
          description: "Not logined"
        "500":
          description: "Error on closing deal"
    delete:
      tags:
        - "deals"
      summary: "Delete a deal"
      description: ""
      operationId: "deleteDeal"
      produces:
        - "application/json"
      parameters:
        - name: "dealId"
          in: "path"
          description: "Deal id to delete"
          required: true
          type: "integer"
      responses:
        "200":
          description: "OK"
        "401":
          description: "Not logined"
        "500":
          description: "Error on deleting deal"
  /api/v1/closed/{closedDealId}:
    get:
      tags:
        - "deals"
      summary: "Get information about one closes deal"
      operationId: "getClosedDeal"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "closedDealId"
          in: "path"
          description: "ID of closed deal"
          required: true
          type: "integer"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/Deals"
        "500":
          description: "Error on getting closed deal"
    delete:
      tags:
        - "deals"
      summary: "Delete a closed deal"
      description: ""
      operationId: "deleteClosedDeal"
      produces:
        - "application/json"
      parameters:
        - name: "closedDealId"
          in: "path"
          description: "Closed deal id to delete"
          required: true
          type: "integer"
      responses:
        "200":
          description: "OK"
        "401":
          description: "Not logined"
        "500":
          description: "Error on deleting closed deal"
definitions:
...
  OnePortfolio:
    type: "object"
    properties:
      name:
        type: "string"
      description:
        type: "string"
      is_public:
        type: "boolean"
  PortfolioDeals:
    type: "object"
    properties:
      PortfolioResp:
        $ref: "#/definitions/OnePortfolio"
      DealResp:
        type: array
        $ref: "#/definitions/Deals"
  Deals:
    type: "object"
    properties:
      id:
        type: "integer"
      ticker:
        type: "string"
      name:
        type: "string"
      type:
        type: "string"
        enum: [bond, stock, fund]
      description:
        type: "string"
      amount:
        type: "integer"
      cost:
        type: "number"
        format: "float"
      value:
        type: "number"
        format: "float"
      currency:
        type: "string"
        enum: [usd, eur, rub]
      opened_at:
        type: "string"
        format: "date-time"
      closed_at:
        type: "string"
        format: "date-time"
      buy_cost:
        type: "number"
        format: "float"
      sell_cost:
        type: "number"
        format: "float"
      profit:
        type: "number"
        format: "float"
      percent:
        type: "number"
        format: "float"
...